+++ libc-bottom-half/cloudlibc/src/libc/sys/uio/pwritev.c
@@ -5,7 +5,7 @@
 #include <sys/types.h>
 #include <sys/uio.h>
 
-#include <wasix/api.h>
+#include <wasi/api.h>
 #include <errno.h>
 
 ssize_t pwritev(int fildes, const struct iovec *iov, int iovcnt, off_t offset) {
@@ -13,7 +13,7 @@ ssize_t pwritev(int fildes, const struct iovec *iov, int iovcnt, off_t offset) {
     errno = EINVAL;
     return -1;
   }
-  uint64_t bytes_written;
+  uint32_t bytes_written;
   __wasi_errno_t error = __wasi_fd_pwrite(
       fildes, (const __wasi_ciovec_t *)iov, iovcnt, offset, &bytes_written);
   if (error != 0) {
