+++ libc-bottom-half/cloudlibc/src/libc/sys/uio/readv.c
@@ -5,7 +5,7 @@
 #include <sys/uio.h>
 
 #include <assert.h>
-#include <wasix/api.h>
+#include <wasi/api.h>
 #include <errno.h>
 #include <stddef.h>
 
@@ -29,7 +29,7 @@ ssize_t readv(int fildes, const struct iovec *iov, int iovcnt) {
     errno = EINVAL;
     return -1;
   }
-  uint64_t bytes_read;
+  uint32_t bytes_read;
   __wasi_errno_t error = __wasi_fd_read(
       fildes, (const __wasi_iovec_t *)iov, iovcnt, &bytes_read);
   if (error != 0) {
