+++ libc-bottom-half/cloudlibc/src/libc/unistd/pread.c
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: BSD-2-Clause
 
-#include <wasix/api.h>
+#include <wasi/api.h>
 #include <errno.h>
 #include <unistd.h>
 
@@ -12,7 +12,7 @@ ssize_t pread(int fildes, void *buf, size_t nbyte, off_t offset) {
     return -1;
   }
   __wasi_iovec_t iov = {.buf = buf, .buf_len = nbyte};
-  uint64_t bytes_read;
+  uint32_t bytes_read;
   __wasi_errno_t error =
       __wasi_fd_pread(fildes, &iov, 1, offset, &bytes_read);
   if (error != 0) {
