+++ libc-bottom-half/cloudlibc/src/libc/unistd/readlinkat.c
@@ -4,14 +4,14 @@
 
 #include <common/errno.h>
 
-#include <wasix/api.h>
+#include <wasi/api.h>
 #include <errno.h>
 #include <string.h>
 #include <unistd.h>
 
 ssize_t __wasilibc_nocwd_readlinkat(int fd, const char *restrict path, char *restrict buf,
                                     size_t bufsize) {
-  uint64_t bufused;
+  uint32_t bufused;
   // TODO: Remove the cast on `buf` once the witx is updated with char8 support.
   __wasi_errno_t error = __wasi_path_readlink(fd, path,
                                                       (uint8_t*)buf, bufsize, &bufused);
