+++ libc-bottom-half/cloudlibc/src/libc/unistd/write.c
@@ -2,13 +2,13 @@
 //
 // SPDX-License-Identifier: BSD-2-Clause
 
-#include <wasix/api.h>
+#include <wasi/api.h>
 #include <errno.h>
 #include <unistd.h>
 
 ssize_t write(int fildes, const void *buf, size_t nbyte) {
   __wasi_ciovec_t iov = {.buf = buf, .buf_len = nbyte};
-  uint64_t bytes_written;
+  uint32_t bytes_written;
   __wasi_errno_t error =
       __wasi_fd_write(fildes, &iov, 1, &bytes_written);
   if (error != 0) {
