+++ libc-bottom-half/sources/__main_void.c
@@ -1,4 +1,4 @@
-#include <wasix/api.h>
+#include <wasi/api.h>
 #include <stdlib.h>
 #include <sysexits.h>
 
@@ -13,8 +13,8 @@ int __main_void(void) {
     __wasi_errno_t err;
 
     // Get the sizes of the arrays we'll have to create to copy in the args.
-    size_t argv_buf_size;
-    size_t argc;
+    uint64_t argv_buf_size;
+    uint32_t argc;
     err = __wasi_args_sizes_get(&argc, &argv_buf_size);
     if (err != __WASI_ERRNO_SUCCESS) {
         _Exit(EX_OSERR);
