+++ libc-bottom-half/sources/__wasilibc_initialize_environ.c
@@ -1,9 +1,9 @@
 #include <unistd.h>
 #include <stdlib.h>
 #include <sysexits.h>
-#include <wasix/api.h>
-#include <wasix/libc.h>
-#include <wasix/libc-environ.h>
+#include <wasi/api.h>
+#include <wasi/libc.h>
+#include <wasi/libc-environ.h>
 
 /// If the program doesn't use `environ`, it'll get this version of
 /// `__wasilibc_environ`, which isn't initialized with a constructor function.
@@ -27,8 +27,8 @@ static char *empty_environ[1] = { NULL };
 // See the comments in libc-environ.h.
 void __wasilibc_initialize_environ(void) {
     // Get the sizes of the arrays we'll have to create to copy in the environment.
-    size_t environ_count;
-    size_t environ_buf_size;
+    uint32_t environ_count;
+    uint64_t environ_buf_size;
     __wasi_errno_t err = __wasi_environ_sizes_get(&environ_count, &environ_buf_size);
     if (err != __WASI_ERRNO_SUCCESS) {
         goto oserr;
